Create credentials from dict
Instead of from a file. Previously the app wrote a file to the system
which could be read back in and used to create credentials.

To be 12 factory it the app shouldn't be writing to the file system like
this. Using the from_json_keyfile_dict method means that no file needs
writing. Nice.